Move legacy stt to separate module
